[BUGFIX] Properly encode file/folder names in the public URL
Properly encoding the file and folder names is required to ensure
proper operation. In the frontend the browsers are very forgiving
and fix a lot of stuff in the background but the TYPO3 backend
checks if it's a valid URL to determine wether it's a local file
that should be referenced relative to the current script or not.

Change-Id: If626584a7a16311e8a303bf1e0014e74eea050b3
Resolves: #66665