[BUGFIX] Using headObject to proof that a folder exists does not work in every case
If files in the bucket were created by a 3rd party tool and the folder structure is not created explicity beforehand (which is not necesssary because S3 allows a structured filesystem just by prefixes), the headObject method returns a 404, because there is no folder object to fetch meta data from. 

We can use listObjects in this case. If it delivers min. 1 object, we can assume that it is a folder.