[BUGFIX] Do not add content when creating folder
When creating a folder with createObject (PUT) and a space " " as content,
the MinIO storage backend breaks and gives errors like
> XMinioObjectExistsAsDirectory
> Object name already exists as a directory.

When passing no content, the folder gets created as normal folder.

The single space was introduced in
> commit 2efda54b61e982b2e923f4f8b1efe6e5117f6bc6
> [FEATURE] Add cache headers to files
>
> Add cache header to files on upload / create and copy files
>
> Resolves: #64001

but I think that was not intentional.

Part of: https://github.com/andersundsehr/aus_driver_amazon_s3/issues/62