[BUGFIX] Use different cache keys for different bucket URLs
When having multiple S3-based storages in TYPO3, the file tree view
shows the same folders for all storages since commit
> 6d01c1abea2741692a5254ce487ed007eaebfebd
> Make some caches configurable, fixes a caching issue
> and minor performance improvements

The culprit is that the caching layer does not take the bucket
configuration into account, leading to the same cache key "/"
for all storages.

This commit uses the S3 client endpoint as prefix for all cache keys.